Rename prase to prales
