+++ settings.gradle
@@ -9,4 +9,4 @@
 
 rootProject.name = 'klang'
 include('klang')
-include('klang-prase')
+include('klang-prales')
